Task 1/requirements.txt
openai
python-dotenv
pytest
pytest-asyncio
httpx
